refactor: migrate default theme to use script-setup
