+++ src/client/theme-default/Layout.vue
@@ -47,7 +47,7 @@
   <Debug />
 </template>
 
-<script>
+<script setup lang="ts">
 import { ref, computed, watch } from 'vue'
 import NavBar from './components/NavBar.vue'
 import Home from './components/Home.vue'
@@ -56,76 +56,56 @@ import SideBar from './components/SideBar.vue'
 import Page from './components/Page.vue'
 import { useRoute, useSiteData, useSiteDataByRoute } from 'vitepress'
 
-export default {
-  components: {
-    Home,
-    NavBar,
-    ToggleSideBarButton,
-    SideBar,
-    Page
-  },
+const route = useRoute()
+const siteData = useSiteData()
+const siteRouteData = useSiteDataByRoute()
 
-  setup() {
-    const route = useRoute()
-    const siteData = useSiteData()
-    const siteRouteData = useSiteDataByRoute()
+const openSideBar = ref(false)
+const enableHome = computed(() => !!route.data.frontmatter.home)
 
-    const openSideBar = ref(false)
-    const enableHome = computed(() => !!route.data.frontmatter.home)
-
-    const showNavbar = computed(() => {
-      const { themeConfig } = siteRouteData.value
-      const { frontmatter } = route.data
-      if (
-        frontmatter.navbar === false
-        || themeConfig.navbar === false) {
-        return false
-      }
-      return (
-        siteData.value.title
-        || themeConfig.logo
-        || themeConfig.repo
-        || themeConfig.nav
-      )
-    })
-
-    const showSidebar = computed(() => {
-      const { frontmatter } = route.data
-      const { themeConfig } = siteRouteData.value
-      return (
-        !frontmatter.home
-        && frontmatter.sidebar !== false
-        && ((typeof themeConfig.sidebar === 'object') && (Object.keys(themeConfig.sidebar).length != 0)
-        || (Array.isArray(themeConfig.sidebar) && themeConfig.sidebar.length != 0))
-      )
-    })
-
-    const pageClasses = computed(() => {
-      return [{
-        'no-navbar': !showNavbar.value,
-        'sidebar-open': openSideBar.value,
-        'no-sidebar': !showSidebar.value
-      }]
-    })
-
-    const toggleSidebar = (to) => {
-      openSideBar.value = typeof to === 'boolean' ? to : !openSideBar.value
-    }
-
-    const hideSidebar = toggleSidebar.bind(null, false)
-    // close the sidebar when navigating to a different location
-    watch(route, hideSidebar)
-    // TODO: route only changes when the pathname changes
-    // listening to hashchange does nothing because it's prevented in router
-
-    return {
-      showNavbar,
-      showSidebar,
-      openSideBar,
-      pageClasses,
-      enableHome,
-      toggleSidebar
-    }
+const showNavbar = computed(() => {
+  const { themeConfig } = siteRouteData.value
+  const { frontmatter } = route.data
+  if (frontmatter.navbar === false || themeConfig.navbar === false) {
+    return false
   }
+  return (
+    siteData.value.title ||
+    themeConfig.logo ||
+    themeConfig.repo ||
+    themeConfig.nav
+  )
+})
+
+const showSidebar = computed(() => {
+  const { frontmatter } = route.data
+  const { themeConfig } = siteRouteData.value
+  return (
+    !frontmatter.home &&
+    frontmatter.sidebar !== false &&
+    ((typeof themeConfig.sidebar === 'object' &&
+      Object.keys(themeConfig.sidebar).length != 0) ||
+      (Array.isArray(themeConfig.sidebar) && themeConfig.sidebar.length != 0))
+  )
+})
+
+const pageClasses = computed(() => {
+  return [
+    {
+      'no-navbar': !showNavbar.value,
+      'sidebar-open': openSideBar.value,
+      'no-sidebar': !showSidebar.value
+    }
+  ]
+})
+
+const toggleSidebar = (to) => {
+  openSideBar.value = typeof to === 'boolean' ? to : !openSideBar.value
 }
+
+const hideSidebar = toggleSidebar.bind(null, false)
+// close the sidebar when navigating to a different location
+watch(route, hideSidebar)
+// TODO: route only changes when the pathname changes
+// listening to hashchange does nothing because it's prevented in router
 </script>
