+++ src/client/theme-default/components/EditLink.vue
@@ -6,25 +6,11 @@
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script setup lang="ts">
 import { useEditLink } from '../composables/editLink'
 import OutboundLink from './icons/OutboundLink.vue'
 
-export default defineComponent({
-  components: {
-    OutboundLink
-  },
-
-  setup() {
-    const { url, text } = useEditLink()
-
-    return {
-      url,
-      text
-    }
-  }
-})
+const { url, text } = useEditLink()
 </script>
 
 <style scoped>
