--- src/client/theme-default/components/NavBar.ts
@@ -1,12 +1,0 @@
-import { withBase } from '../utils'
-import NavBarLinks from './NavBarLinks.vue'
-
-export default {
-  components: {
-    NavBarLinks
-  },
-
-  setup() {
-    return { withBase }
-  }
-}
