+++ src/client/theme-default/components/NavBarLink.vue
@@ -14,7 +14,67 @@
   </div>
 </template>
 
-<script src="./NavBarLink"></script>
+<script setup lang="ts">
+import { computed, defineOptions } from 'vue'
+import { useRoute } from 'vitepress'
+import { withBase, isExternal } from '../utils'
+import type { DefaultTheme } from '../config'
+import OutboundLink from './icons/OutboundLink.vue'
+
+const normalizePath = (path: string): string => {
+  path = path
+    .replace(/#.*$/, '')
+    .replace(/\?.*$/, '')
+    .replace(/\.html$/, '')
+  if (path.endsWith('/')) {
+    path += 'index'
+  }
+  return path
+}
+
+const { props } = defineOptions<{
+  props: {
+    item: DefaultTheme.NavItemWithLink
+  }
+}>()
+
+const item = props.item
+
+const route = useRoute()
+
+const classes = computed(() => ({
+  active: isActiveLink.value,
+  external: isExternalLink.value
+}))
+
+const isActiveLink = computed(() => {
+  return normalizePath(withBase(item.link)) === normalizePath(route.path)
+})
+
+const isExternalLink = computed(() => {
+  return isExternal(item.link)
+})
+
+const href = computed(() => {
+  return isExternalLink.value ? item.link : withBase(item.link)
+})
+
+const target = computed(() => {
+  if (item.target) {
+    return item.target
+  }
+
+  return isExternalLink.value ? '_blank' : ''
+})
+
+const rel = computed(() => {
+  if (item.rel) {
+    return item.rel
+  }
+
+  return isExternalLink.value ? 'noopener noreferrer' : ''
+})
+</script>
 
 <style>
 .nav-item {
