+++ src/client/theme-default/components/Page.vue
@@ -13,17 +13,9 @@
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
+<script setup lang="ts">
 import EditLink from './EditLink.vue'
 import NextAndPrevLinks from './NextAndPrevLinks.vue'
-
-export default defineComponent({
-  components: {
-    EditLink,
-    NextAndPrevLinks
-  }
-})
 </script>
 
 <style scoped>
